perf: ttfb improvements
