+++ apps/web/app/middleware/pwa-update.global.client.ts
