+++ apps/web/app/pages/index.vue
@@ -18,7 +18,7 @@ definePageMeta({
   isBlockified: true,
   type: 'immutable',
   identifier: 'index',
-  middleware: ['newsletter-confirmation-client', 'notify-me-interactions'],
+  middleware: ['newsletter-confirmation-client', 'notifyme-interactions-client'],
 });
 
 const useLocaleSpecificHomepageTemplate = (locale: string) =>
